+++ GoGreen/Controllers/EventController.cs
@@ -78,17 +78,11 @@ namespace GoGreen.Controllers
         }
 
         // POST: api/Event
-        [AllowAnonymous]
+        [Authorize]
         [HttpPost]
         [Consumes("multipart/form-data")]
         public async Task<ActionResult<EventResponse>> PostEvent([FromForm] EventRequest request, IFormFile? imageFile)
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
-            if (string.IsNullOrEmpty(userId))
-            {
-                return BadRequest("The user ID claim is missing");
-            }
 
             var type = await _context.EventTypes.FindAsync(request.TypeId);
 
@@ -138,11 +132,10 @@ namespace GoGreen.Controllers
         [HttpPut("{id}")]
         public async Task<ActionResult<EventResponse>> PutEvent(int id, [FromBody] EventRequest request)
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
-            if (string.IsNullOrEmpty(userId))
+            if (!await CheckPermisionAsync(id))
             {
-                return BadRequest("The user ID claim is missing");
+                return BadRequest("The user has no permission for this action");
             }
 
             var updatedEvent = await _eventService.Update(id, request);
@@ -152,10 +145,15 @@ namespace GoGreen.Controllers
         }
 
         // DELETE: api/Event/5
-        [Authorize(Roles = "super-admin")]
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteEvent(int id)
         {
+            
+            if (!await CheckPermisionAsync(id))
+            {
+                return BadRequest("The user has no permission for this action");
+            }
 
             var isDeleted = await _eventService.Delete(id);
 
@@ -173,15 +171,13 @@ namespace GoGreen.Controllers
         public async Task<ActionResult<EventResponse>> AddEventImage(int eventId, IFormFile imageFile)
 
         {
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-
-            if (string.IsNullOrEmpty(userId))
+            if (!await CheckPermisionAsync(eventId))
             {
-                return BadRequest("The user ID claim is missing");
+                return BadRequest("The user has no permission for this action");
             }
 
             var @event = await _context.Events
-                .Where(e => e.Id == eventId && e.UserId == userId)
+                .Where(e => e.Id == eventId)
                 .Include(a => a.EventImages)
                     .ThenInclude(ei => ei.Image)
                 .FirstOrDefaultAsync();
@@ -223,16 +219,14 @@ namespace GoGreen.Controllers
         [HttpDelete("{eventId}/Image/{imageId}")]
         public async Task<IActionResult> DeleteEventImage(int eventId, int imageId )
         {
-            
-            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
-            if (string.IsNullOrEmpty(userId))
+            if (!await CheckPermisionAsync(eventId))
             {
-                return BadRequest("The user ID claim is missing");
+                return BadRequest("The user has no permission for this action");
             }
 
             var @event = await _context.Events
-                .Where(e => e.Id == eventId && e.UserId == userId)
+                .Where(e => e.Id == eventId)
                 .Include(e => e.EventImages)
                     .ThenInclude(ei => ei.Image)
                 .FirstOrDefaultAsync();
@@ -268,6 +262,30 @@ namespace GoGreen.Controllers
             return Ok(recommendations);
         }
 
+        private async Task<bool> CheckPermisionAsync(int id)
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+
+            if (string.IsNullOrEmpty(userId))
+            {
+                return false;
+            }
+
+            var user = await _context.User.Include(e => e.Municipality).FirstOrDefaultAsync(u => u.Id == userId);
+            var data = await _context.Events
+                        .Where(a => a.MunicipalityId == user.MunicipalityId)
+                        .Where(a => a.Id == id)
+                        .FirstOrDefaultAsync();
+
+            if (data == null)
+            {
+                return false;
+            }
+
+            return true;
+
+        }
 
     }
 }
